+++ packages/typir/src/graph/type-graph.ts
@@ -39,7 +39,7 @@ export class TypeGraph {
             if (this.nodes.get(mapKey) === type) {
                 // this type is already registered => that is OK
             } else {
-                throw new Error(`Names of types must be unique: ${mapKey}`);
+                throw new Error(`There is already a type with the identifier '${mapKey}'.`);
             }
         } else {
             this.nodes.set(mapKey, type);